+++ week4/README.md
@@ -3,3 +3,15 @@
 //% 구현 과정에서 예외처리 확실하게 할 것(사실 inputNumber를 스택에 바로바로 넣었다 뺐다 하면 해결)
 // 전체적인 구현 로직 다시 한 번 생각해볼 것
 -> 현재는 연산자가 들어오면 숫자를 stack에 넣는 방식이지만, 연산자가 들어오기 전에 스택에 숫자를 바로바로 갱신해주는 과정으로 코드를 짜면 보다 전체적인 코드가 간결해질 것 같음.(stack을 적극적으로 활용하자.)
+
+언래핑 강제로 하는거 줄이기.
+스택의 기능 분리(calculator에서 분리하고, stack타입으로 해서 구조체로 해결하기. ))
+스택을 스택 타입으로Stack ; 
+[String]
+Stack ; [String]
+
+
+
+연산용 모델을 따로 분리해서 처리하기.
+
+차라리 let symbol을 스태틱처리 해보기.(타입에 고정된 것이라서)
