+++ week5/Calculator/Calculator/Extensions/Numeric+Extension.swift
@@ -1,0 +1,14 @@
+//
+//  Numeric+Extension.swift
+//  Calculator
+//
+//  Created by ohhyeongseok on 2022/08/25.
+//
+
+import Foundation
+
+extension Numeric {
+    var exponentialNotation: String {
+        return Formatter.scientific.string(for: self) ?? ""
+    }
+}
