+++ week5/Calculator/Calculator/View/ScreenView.swift
@@ -17,8 +17,8 @@ struct ScreenView: View {
                 .lineLimit(1)
                 .foregroundColor(.white)
                 .frame(alignment: .bottomTrailing)
-                .font(.system(size: 50))
-                .minimumScaleFactor(0.3)
+                .font(.system(size: 80))
+                .minimumScaleFactor(0.5)
                 .padding()
         }
     }
